drfdockerized/requirements.txt
Django==5.1.3
asgiref==3.8.1
djangorestframework==3.15.0
gunicorn==21.2.0
pygments==2.14.0